% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr2.R
\name{\%notin}
\alias{\%notin}
\title{While converting the names, BiTr usually returns a df with two columns, "fromType" & "toType", and it is oftenly difficult to match them
to the original dataset. This is a simple function around that concept that BiTr should bind to converted IDs back to the original df
for easier downstream analysis
Dependencies:
This package depends on the ClusterProfiler (4.4.4) function of BiTR, that can
be installed using BiocManager::install("clusterProfiler")
and package for Genome wide annotation for human (default, "org.Hs.eg.db"),
and can be installed using BiocManager::install("org.Hs.eg.db")}
\usage{
`\%notin`(...)
}
\arguments{
\item{input}{an R object/data frame that can be coercible to df}

\item{colName}{string; column name containing the ids.}

\item{fromType}{string; the name of column that has to be converted; it must be one of types supported by BiTR}

\item{toType}{String; desired output; it must be one of types supported by BiTR}

\item{rm.duplicates}{Boolean; Duplicated names to be removed or not.}

\item{moveToFirst}{if True, the new column will be moved to be the first column in the new df; default is FALSE}
}
\description{
While converting the names, BiTr usually returns a df with two columns, "fromType" & "toType", and it is oftenly difficult to match them
to the original dataset. This is a simple function around that concept that BiTr should bind to converted IDs back to the original df
for easier downstream analysis
Dependencies:
This package depends on the ClusterProfiler (4.4.4) function of BiTR, that can
be installed using BiocManager::install("clusterProfiler")
and package for Genome wide annotation for human (default, "org.Hs.eg.db"),
and can be installed using BiocManager::install("org.Hs.eg.db")
}
