% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr2.R
\name{matchID}
\alias{matchID}
\title{The following function was an attempt to identify the ID name, however, it is
reasonably difficult to guess the correct name, in case the provided ID is incorrect.}
\usage{
matchID(unknown)
}
\description{
The following function was an attempt to identify the ID name, however, it is
reasonably difficult to guess the correct name, in case the provided ID is incorrect.
}
